+++ tests/test_groups.py
@@ -6,8 +6,10 @@ import pytest
 from conftest import utc_offset
 from homematicip.group import *
 from homematicip.home import Home
-from homematicip_demo.helper import (fake_home_download_configuration,
-                                     no_ssl_verification)
+from homematicip_demo.helper import (
+    fake_home_download_configuration,
+    no_ssl_verification,
+)
 
 
 def test_meta_group(fake_home: Home):