+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- CLI
+  - --channel parameter for turning on/off different channels and not just the first one
+
 ## [0.10.17] 2020-02-16
 
 ### FIXED